+++ source/Roy.Logging/Roy.Logging/Helpers/RecordService.cs
@@ -1,14 +1,13 @@
 ﻿using Avalon.Base.Extension.Collections;
 using Roy.Logging.Domain.Attributes;
 using Roy.Logging.Domain.Settings.Attributes;
-using Roy.Logging.Helpers;
 
-namespace Roy.Logging.Services;
+namespace Roy.Logging.Helpers;
 
 /// <summary>
-/// Register service.
+/// Record register service.
 /// </summary>
-internal class Record
+internal class RecordService
 {
     /// <summary>
     /// Save message.
@@ -31,12 +30,12 @@ internal class Record
         catch { }
 
         try
-        { 
+        {
             if (setting.Databases.HasElements())
             {
                 new DatabaseService().SaveAsync(message, setting.Databases);
             }
-        } 
+        }
         catch { }
 
         try
