+++ source/Roy.Logging/Roy.Logging/LoggerExtension.cs
@@ -4,7 +4,7 @@ using Roy.Logging.Domain.Attributes;
 using Roy.Logging.Domain.Contants;
 using Roy.Logging.Domain.Program;
 using Roy.Logging.Domain.Settings;
-using Roy.Logging.Services;
+using Roy.Logging.Helpers;
 using System.Diagnostics;
 
 namespace Roy.Logging;
@@ -209,7 +209,7 @@ public static class LoggerExtension
                     frame, setting.Log.LoadInformationSettings, webApplicationHttpContext,
                     listOfParameters);
 
-                new Record().SaveAsync(
+                new RecordService().SaveAsync(
                     detail, setting.Log);
             }
         }
